Consolidated ExtensionTest objects and minor refactoring
